Completing app refactoring with external components
