+++ apps/cms/next-env.d.ts
@@ -1,5 +1,6 @@
-/// <reference types="next" />
-/// <reference types="next/image-types/global" />
+// / <reference types="next" />
+// / <reference types="next/image-types/global" />
+// / <reference types="next/navigation-types/compat/navigation" />
 
 // NOTE: This file should not be edited
 // see https://nextjs.org/docs/app/building-your-application/configuring/typescript for more information.
