+++ apps/cms/src/app/(public)/auth/[authAction]/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 import "@tanam/cms/assets/scss/layout-authentication.scss";
-import {useFirebaseUi} from "@tanam/cms/hooks/useFirebaseUi";
 import Image from "next/image";
 import {notFound, useParams} from "next/navigation";
 import {useEffect} from "react";
+import {useFirebaseUi} from "../../../../../hooks/useFirebaseUi";
 
 export default function AuthPage() {
   const {authAction} = useParams<{authAction: string}>() ?? {};
