+++ apps/cms/src/components/Auth/PublicOnlyRestriction.tsx
@@ -1,5 +1,5 @@
 "use client";
-import {useAuthentication} from "../../../../../hooks/useAuthentication";
+import {useAuthentication} from "../../hooks/useAuthentication";
 
 interface PublicOnlyRestrictionProps {
   children: React.ReactNode;
