+++ apps/cms/src/components/DarkModeSwitcher.tsx
@@ -1,8 +1,8 @@
 import {Switcher} from "@tanam/ui-components";
 import {useEffect} from "react";
-import {useAuthentication} from "../../../../../hooks/useAuthentication";
-import useColorMode from "../../../../../hooks/useColorMode";
-import {useTanamUser} from "../../../../../hooks/useTanamUser";
+import {useAuthentication} from "../hooks/useAuthentication";
+import useColorMode from "../hooks/useColorMode";
+import {useTanamUser} from "../hooks/useTanamUser";
 
 const DarkModeSwitcher = () => {
   const [colorMode, setColorMode] = useColorMode();
@@ -13,7 +13,7 @@ const DarkModeSwitcher = () => {
     if (typeof setColorMode === "function" && !!tanamUser?.colorMode) {
       setColorMode(tanamUser.colorMode);
     }
-  }, [tanamUser]);
+  }, [setColorMode, tanamUser]);
 
   return (
     <Switcher
