+++ apps/cms/src/hooks/useAuthentication.tsx
@@ -1,9 +1,9 @@
 "use client";
-import {firebaseAuth} from "@tanam/cms/plugins/firebase";
-import {TanamRole} from "@tanam/domain-shared";
+import {TanamRole} from "@tanam/domain-client";
 import {User} from "firebase/auth";
 import {redirect, usePathname} from "next/navigation";
 import {useEffect, useState} from "react";
+import {firebaseAuth} from "../plugins/firebase";
 
 export function useAuthentication() {
   const pathname = usePathname();
