+++ apps/cms/src/hooks/useGenkitArticle.tsx
@@ -1,11 +1,9 @@
-import {TanamDocumentClient} from "@tanam/cms/models/TanamDocumentClient";
-import {UserNotification} from "@tanam/cms/models/UserNotification";
-import {firestore, storage} from "@tanam/cms/plugins/firebase";
-import {generateArticle} from "@tanam/cms/plugins/genkit/article";
-import {ArticleSchema, TanamPublishStatus} from "@tanam/domain-shared";
+import {ArticleSchema, TanamDocumentClient, TanamPublishStatus, UserNotification} from "@tanam/domain-client";
 import {collection, doc, getDocs, limit, orderBy, query, setDoc, where} from "firebase/firestore";
 import {getDownloadURL, ref, uploadBytes} from "firebase/storage";
 import {useEffect, useState} from "react";
+import {firestore, storage} from "../plugins/firebase";
+import {generateArticle} from "../plugins/genkit/article";
 
 // NOTE: (Dennis) For some reason, the enum values are warning for not being used
 export enum ProcessingState {
