+++ apps/cms/src/hooks/useTanamUser.tsx
@@ -1,9 +1,8 @@
-import {TanamUserClient} from "@tanam/cms/models/TanamUserClient";
-import {UserNotification} from "@tanam/cms/models/UserNotification";
-import {firestore, storage} from "@tanam/cms/plugins/firebase";
+import {TanamUserClient, UserNotification} from "@tanam/domain-client";
 import {doc, onSnapshot, updateDoc} from "firebase/firestore";
 import {getDownloadURL, ref} from "firebase/storage";
 import {useEffect, useState} from "react";
+import {firestore, storage} from "../plugins/firebase";
 
 /**
  * Hook to get a Tanam user document from Firestore
